Extract ContainerAvatar to a separate component
The ContainerAvatar logic and component have been extracted from the Services page into a separate, reusable component. This change enhances code readability and promotes component reuse, adhering to best practices for React design patterns.
